agents/platform_mcp_client/requirements.txt
google-cloud-aiplatform[adk,agent_engines]>=1.90.0
Flask
mcp[cli]
google-adk
python-dateutil
humanize
nest_asyncio
asyncclick